listening-comp/backend/requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Data Processing
numpy==1.26.3
pandas==2.1.1  # Standardized pandas version

# ML and AI
openai>=1.0.0
openai-whisper>=20231117
torch==2.1.0  # Required for Whisper
torchaudio==2.1.0  # Required for audio processing
sentence-transformers>=2.2.0

# Database
sqlalchemy>=1.4.41
aiosqlite>=0.19.0

# API and HTTP
httpx==0.24.1
aiohttp==3.8.5
python-multipart>=0.0.5
flask==3.0.0
requests>=2.25.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0
jsonschema==4.20.0

# Configuration
python-dotenv==1.0.0
pydantic>=1.8.2

# Language Processing
langdetect==1.0.9
youtube-transcript-api==0.6.2  # Keep this version consistent

# TTS Services
boto3==1.34.22
google-cloud-texttospeech==2.16.3
google-auth==2.27.0
soundfile==0.12.1
azure-cognitiveservices-speech>=1.19.0

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0
SpeechRecognition==3.10.0
sounddevice>=0.4.5
librosa>=0.10.0

# Frontend
streamlit==1.33.0  # Standardized to match frontend version
streamlit-webrtc==0.47.1
streamlit-audiorecorder==0.0.3  # Updated to match frontend version

# Japanese language processing
fugashi>=1.1.0
unidic-lite==1.0.8  # Lighter version of unidic
mecab-python3==1.0.8  # Required by fugashi
ipadic>=1.0.0  # Japanese dictionary
romkan>=0.2.1
pykakasi>=2.2.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2

# Rate limiting for production
redis==5.0.1
Flask-Limiter[redis]==3.5.0

# Backend-specific
gunicorn==21.2.0

typing-extensions>=4.5.0  # Added to ensure compatibility

# Utilities
tqdm>=4.64.0
